refactor: rename nc package to node_classification
