--- docs/api/nc.md
@@ -1,3 +1,0 @@
-# Node classification (NC)
-
-::: hypertorch.nc
