+++ docs/api/node_classification.md
@@ -1,0 +1,3 @@
+# Node classification (NC)
+
+::: hypertorch.node_classification
