+++ examples/node_classification/node2vec.py
@@ -2,7 +2,7 @@ from torchmetrics import MetricCollection
 from torchmetrics.classification import MulticlassAUROC, MulticlassAccuracy, MulticlassF1Score
 
 from hypertorch.data import AlgebraDataset, DataLoader, Node2VecEnricher
-from hypertorch.nc import Node2VecClassifier
+from hypertorch.node_classification import Node2VecClassifier
 from hypertorch.train import MultiModelTrainer
 from hypertorch.types import ModelConfig
 from hypertorch.utils import node_labels_from_node_degrees
