+++ hypertorch/integration_tests/nc/hnhn_nc_integration_test.py
@@ -10,7 +10,7 @@ from hypertorch.integration_tests.common import (
     split_dataset,
     train_test_loop,
 )
-from hypertorch.nc import HNHNClassifier
+from hypertorch.node_classification import HNHNClassifier
 
 
 NUM_CLASSES = 3
