+++ hypertorch/integration_tests/nc/hypergcn_nc_integration_test.py
@@ -11,7 +11,7 @@ from hypertorch.integration_tests.common import (
     split_dataset,
     train_test_loop,
 )
-from hypertorch.nc import HyperGCNClassifier
+from hypertorch.node_classification import HyperGCNClassifier
 
 
 NUM_CLASSES = 3
