+++ hypertorch/integration_tests/nc/mlp_nc_integration_test.py
@@ -10,7 +10,7 @@ from hypertorch.integration_tests.common import (
     split_dataset,
     train_test_loop,
 )
-from hypertorch.nc import MLPClassifier
+from hypertorch.node_classification import MLPClassifier
 
 
 NUM_CLASSES = 3
