+++ hypertorch/integration_tests/nc/node2vec_nc_integration_test.py
@@ -9,7 +9,7 @@ from hypertorch.integration_tests.common import (
     split_dataset,
     train_test_loop,
 )
-from hypertorch.nc import Node2VecClassifier
+from hypertorch.node_classification import Node2VecClassifier
 from hypertorch.types import TaskEnum
